Add files via upload
The pdf file is the first research paper I wrote in development economics. It investigates the impact of corruption on deforestation.
The do file is the Stata coding file that displays all the codes for cleaning, and manipulating the dataset, and further displays regression analysis using the panel fixed effects method.